requirements.txt
pillow==11.3.0
tqdm==4.67.1
